fix: リダイレクトが親プロセスの fd を壊す問題を修正
外部コマンドのリダイレクトは子プロセスのみで実行し、
ビルトインのリダイレクトは dup で fd を保存・復元するよう変更。
cmd_not_found と exec_external_cmd を find_exec_path.c に移動。
